backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.2.0
sqlalchemy
psycopg2-binary
pillow
pydub
moviepy
python-dotenv
img2pdf
pdf2image
